requirements.txt
fastapi
uvicorn
hypercorn
psycopg2-binary
python-dotenv
pydantic
gunicorn
requests